+++ src/cli/printers/mod.rs
@@ -1,4 +1,5 @@
 mod debug;
+mod json;
 mod pretty;
 mod print_style;
 mod ser;